pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecommerce-audit"
version = "0.1.0"
description = "Multimodal e-commerce product auditing with Qwen3-VL and VeRL"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Ecommerce Audit Team" }]
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "pandas>=2.2",
  "pyarrow>=16.0",
  "Pillow>=10.0",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "httpx>=0.27",
  "transformers>=4.42",
  "torch>=2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_configs = true
